% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makeMu}
\alias{makeMu}
\title{Make mu matrix}
\usage{
makeMu(tau_pos, tau_vel, dt)
}
\arguments{
\item{tau_pos}{Parameter \eqn{tau[pos]} of the movement process}

\item{tau_vel}{Parameter \eqn{tau[vel]} of the movement process}

\item{dt}{Time interval}
}
\value{
Green's Function matrix
}
\description{
This code is adapted from the langevin function in the R package ctmm (Calabrese et al., 2016).
It is an R wrapper for the internal makeT function
}
\references{
Calabrese, J.M., Fleming, C.H. and Gurarie, E. (2016).
ctmm: an r package for analyzing animal relocation data as a continuous‐time stochastic process.
Methods Ecol Evol, 7: 1124-1132. doi:10.1111/2041-210X.12559
}
