% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulate from 2D IOU/OUF process}
\usage{
sim(obsTimes, tau_pos, tau_vel, sigma, Q)
}
\arguments{
\item{obsTimes}{Vector of observation times}

\item{tau_pos}{Parameter \eqn{tau[pos]} of the movement process}

\item{tau_vel}{Parameter \eqn{tau[vel]} of the movement process}

\item{sigma}{Parameter \eqn{sigma} of the movement process}

\item{Q}{Infinitesimal generator matrix}
}
\value{
Simulated track
}
\description{
Simulate from 2D IOU/OUF process
}
