% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateState.R
\name{updateState}
\alias{updateState}
\title{Update state sequence}
\usage{
updateState(
  obs,
  nbStates,
  knownStates,
  switch,
  updateLim,
  param,
  mu,
  Hmat,
  updateProbs = NULL,
  rateparam = NULL,
  kappa = NULL,
  model = NULL
)
}
\arguments{
\item{obs}{Matrix of observations, with columns \code{x}, \code{y}, \code{time}, and \code{state}}

\item{knownStates}{Vector of known, fixed states}

\item{switch}{Matrix of state switches, with columns \code{time} and \code{state}}

\item{updateLim}{Vector of two elements: min and max length of updated interval
(each an integer, corresponding to a number of observations)}

\item{updateProbs}{Vector of probabilities of picking each number from
updateLim\link{1} to updateLim\link{2}, for the length of the update interval}

\item{kappa}{Upper bounds of transition rate}
}
\value{
List of two elements:
\itemize{
\item newSwitch: Updated matrix of transitions
\item newData: Updated matrix of all data (observations and transitions)
}
}
\description{
Update state sequence
}
\details{
The update is done with the function \code{sample_path} from the
package ECctmc (Fintzi, 2017).
}
\references{
Jon Fintzi (2017). ECctmc: Simulation from Endpoint-Conditioned Continuous
Time Markov Chains. R package version 0.2.4.
https://CRAN.R-project.org/package=ECctmc

TODO: needs priors
}
