% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCMC.R
\name{runMCMC}
\alias{runMCMC}
\title{Run MCMC iterations}
\usage{
runMCMC(
  track,
  nbStates,
  nbIter,
  fixpar = NULL,
  inits,
  priors,
  props,
  tunes,
  Hmat,
  updateState = TRUE,
  mc.cores = 1
)
}
\arguments{
\item{track}{Dataframe of data, with columns "x", "y", "time", and "ID"}

\item{nbStates}{Number of states}

\item{nbIter}{Number of iterations}

\item{fixpar}{Vector of fixed parameter values (tau_pos, tau_vel, sigma), with NA for parameters to estimate}

\item{inits}{List of initial parameters
(tau_pos, tau_vel, sigma, Q, state)}

\item{priors}{List of parameters of prior distributions, with components:
\itemize{
  \item{"mean":} Vector of means for normal priors on movement parameters, of length 3*nbStates
  \item{"sd":} Vector of standard deviations for normal priors on movement parameters, of length
3*nbStates
  \item{"shape":} Vector of shapes of gamma priors for the transition rates
  \item{"rate":} Vector of rates of gamma priors for the transition rates
  \item{"con":} Vector of concentrations of Dirichlet priors for transition probabilities
}}

\item{props}{List of parameters of proposal distributions, with components:
\itemize{
  \item{"tau_posSD":} Scalar standard deviation for normal proposal distribution of tau_pos
  \item{"tau_velSD":} Scalar standard deviation for normal proposal distribution of tau_vel
  \item{"sigmaSD":} Scalar standard deviation for normal proposal distribution of sigma
  \item{"updateLim":} Vector of two values: min and max length of updated state sequence
  \item{"updateProbs":} Probability for each element of updateLim[1]:updateLim[2] (if NULL,
  all values are equiprobable)
}}

\item{tunes}{List of tuning parameters, with components:
\itemize{
  \item{"thinStates":} Thinning factor for the posterior state sequences (needed because
of memory limitations)
}}

\item{Hmat}{Matrix of observation error variance (four columns, and one row
for each row of data)}

\item{updateState}{Logical. If FALSE, the state process is not updated
(for exploratory analysis only, useful for testing single state models)}
}
\description{
Run MCMC iterations
}
